+++ src/Symfony/Component/Validator/Constraints/NotBlankValidator.php
@@ -17,6 +17,7 @@ use Symfony\Component\Validator\Exception\UnexpectedTypeException;
 
 /**
  * @author Bernhard Schussek <bschussek@gmail.com>
+ * @author Kévin Dunglas <dunglas@gmail.com>
  */
 class NotBlankValidator extends ConstraintValidator
 {
@@ -27,6 +28,10 @@ class NotBlankValidator extends ConstraintValidator
     {
         if (!$constraint instanceof NotBlank) {
             throw new UnexpectedTypeException($constraint, __NAMESPACE__.'\NotBlank');
+        }
+
+        if ($constraint->allowNull && null === $value) {
+            return;
         }
 
         if (false === $value || (empty($value) && '0' != $value)) {